+++ PasswordReset.php
@@ -370,7 +370,7 @@ function _sendPasswordResetEmail( $user_id, $user_type = 'staff', $email )
 	$hash = encrypt_password( $user_id . $username . $name . $password . $email . $last_login );
 
 	// Generate link.
-	$link = _currentPageURL() . '?h=' . $hash;
+	$link = RosarioURL( 'script' ) . '?h=' . $hash;
 
 	// Send email.
 	require_once 'ProgramFunctions/SendEmail.fnc.php';
@@ -404,32 +404,6 @@ function _sendPasswordResetEmail( $user_id, $user_type = 'staff', $email )
 	}
 
 	return true;
-}
-
-
-function _currentPageURL()
-{
-	$page_url = 'http';
-
-	if ( isset( $_SERVER['HTTPS'] )
-		&& $_SERVER['HTTPS'] == 'on' )
-	{
-		$page_url .= 's';
-	}
-
-	$page_url .= '://';
-
-	if ( $_SERVER['SERVER_PORT'] != '80'
-		&& $_SERVER['SERVER_PORT'] != '443' )
-	{
-		$page_url .= $_SERVER['SERVER_NAME'] . ':' . $_SERVER['SERVER_PORT'] . $_SERVER['REQUEST_URI'];
-	}
-	else
-	{
-		$page_url .= $_SERVER['SERVER_NAME'] . $_SERVER['REQUEST_URI'];
-	}
-
-	return $page_url;
 }